Update LED() to display new iteration in new color
Before this change, LED would call pixels.clear() to turn of all LED's and to start a new iteartion. In this update, when the full LED ring is blue, the second iteration will overwrite the previous LED state to a green LED. We thought it would be nicer and clearer representation of the time that has passed.